+++ app/board/deprecated.py
@@ -28,7 +28,7 @@ class Wheel:
 
         self.last_pressed = False
         self.pressed = False
-        self.longPressTimer = Timer()
+        self.longPressTimer: Timer | None = None
 
         self.clk.irq(lambda p:self._rotated())
         self.sw.irq(lambda p:self._pressed())
@@ -48,7 +48,7 @@ class Wheel:
             self.timer.deinit()
             self.timer = None
 
-        self.timer = Timer(mode=Timer.ONE_SHOT, period=1000, callback=lambda t: self._led_off())
+        self.timer = Timer(-1, mode=Timer.ONE_SHOT, period=1000, callback=lambda t: self._led_off())
 
 
     def _rotated(self):
@@ -91,7 +91,7 @@ class Wheel:
                 def lpc(t):
                     self._long_action()
 
-                self.longPressTimer.init(mode=Timer.ONE_SHOT, period=PushButton.longpress_ms, callback=lpc)
+                self.longPressTimer = Timer(-1, mode=Timer.ONE_SHOT, period=PushButton.longpress_ms, callback=lpc)
                 self._send_hook()
 
     def _long_action(self):
@@ -99,7 +99,9 @@ class Wheel:
             self._send_hook(True)
 
     def _send_hook(self, long=False):
-        self.timer.deinit()
+        if self.timer is not None:
+            self.timer.deinit()
+            
         self._flash_color()
 
         if long:
